+++ beekeeper-integration-tests/src/test/java/com/expediagroup/beekeeper/integration/BeekeeperPathSchedulerApiaryIntegrationTest.java
@@ -464,7 +464,7 @@ class BeekeeperPathSchedulerApiaryIntegrationTest {
     MeterRegistry meterRegistry = BeekeeperPathSchedulerApiary.meterRegistry();
     List<Meter> meters = meterRegistry.getMeters();
     assertThat(meters).extracting("id", Meter.Id.class).extracting("name")
-      .contains("scheduled-paths");
+      .contains("paths-scheduled");
   }
 
   private SendMessageRequest sendMessageRequest(String payload) {